requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Machine Learning & Deep Learning
tensorflow==2.18.0
transformers==4.36.2
torch==2.5.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# LangChain & OpenAI - Agentes IA
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.10.0
tiktoken==0.5.2

# LLM Providers alternativos (sin depender de OpenAI)
# Ollama: LLM local gratuito (requiere ollama instalado: https://ollama.ai)
# Groq: API gratuita ultrarrápida (console.groq.com)
# HuggingFace: Modelos open-source (huggingface.co)

# Database Drivers
clickhouse-driver==0.2.6
pymongo==4.6.1
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# API & HTTP
httpx==0.26.0
requests==2.31.0

# Data Processing
python-dateutil==2.8.2
pytz==2024.1

# Visualization (para generar reportes)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Monitoring & Logging
python-json-logger==2.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Environment
python-dotenv==1.0.0
